# Add `rewrite_call_expression` through `_rewrite_expr`.