--- a/test/macros.jl
+++ b/test/macros.jl
@@ -155,6 +155,7 @@ function build_constraint_keyword_test(ModelType::Type{<:JuMP.AbstractModel})
     end
 end
 
+<<<<<<< HEAD
 struct CustomType
 end
 function JuMP.parse_constraint_head(_error::Function, ::Val{:(:=)}, lhs, rhs)
@@ -176,6 +177,7 @@ function custom_expression_test(ModelType::Type{<:JuMP.AbstractModel})
     end
 end
 
+<<<<<<< HEAD
 function JuMP.parse_one_operator_constraint(_error::Function, ::Bool, ::Val{:f}, x)
     return :(), :(build_constraint($_error, $(esc(x)), $(esc(CustomType()))))
 end
@@ -191,23 +193,13 @@ function custom_function_test(ModelType::Type{<:JuMP.AbstractModel})
         @test_macro_throws ErrorException @constraint(model, g(x))
     end
 end
+=======
+=======
+JuMP.expression_to_rewrite(head::Val{:donothing}, var) = true
+>>>>>>> 00c7836c... Make the tests pass.
+>>>>>>> e8726d59... Make the tests pass.
 function JuMP.rewrite_call_expression(errorf::Function, head::Val{:donothing}, var)
-    m = gensym()
-    vi = gensym()
-    nvar = gensym()
-
-    parse_code_var = quote
-      $m = owner_model($(esc(index)))
-      $vi = VariableInfo(false, NaN, false, NaN, false, NaN, false, NaN, false, false)
-      $nvar = add_variable($m, build_variable($errorf, $vi), "")
-    end
-
-    idx, parse_code_index = JuMP._MA.rewrite(index)
-    build_code_con = quote
-      add_constraint($m, build_constraint($errorf, _MA.mutable_operate!(-, $var, $nvar), MOI.EqualTo(0.0)))
-    end
-
-    return :($parse_code_var; $parse_code_index), build_code_con, var
+    return :(), :(), esc(var)
 end
 function build_constraint_test(ModelType::Type{<:JuMP.AbstractModel})
     @testset "Extension of @constraint with rewrite_call_expression #2229" begin
@@ -415,6 +407,7 @@ function macros_test(ModelType::Type{<:JuMP.AbstractModel}, VariableRefType::Typ
     build_constraint_keyword_test(ModelType)
     custom_expression_test(ModelType)
     custom_function_test(ModelType)
+    build_constraint_test(ModelType)
 end
 
 @testset "Macros for JuMP.Model" begin