Excerpt from #2051. New take on #2229, much more general (and not much longer — unless you take https://github.com/JuliaOpt/JuMP.jl/pull/2229/commits/fdef385b6b5de779208124f991c9a221f74ae925, which unfortunately does not work…). 

I'd prefer to see this PR merged instead of #2229, due to the functionalities that are possible with this, but maybe this is not the best way to implement it… 

The general goal is to allow rewriting expressions at the JuMP level, only based on syntactic information, to provide more convenience for users. This only provides more points for extensions to plug into JuMP. 

More in details (based on the current developments of the CP sets): the modelling layer can decide to rewrite an expression like `y <= count(x .== 5)` as two constraints, a CP one (say `z == count(x .== 5))`, and a purely linear one (say `y <= z`), all of this at JuMP layer. These new constraints can then be bridged by MOI if needed. The constraints to model in the first hand (like `y <= count(x .== 5))`) should not have MOI equivalents (because that would create a very very large number of sets to support: instead of just Count, you would need LessThanCount, maybe AffineLessThanCount later on, and so on). 

This thus raises the issue of deleting the new constraints: as a user, it only makes sense to have a reference to the constraint `y <= count(x .== 5)`; to delete it, two constraints and a variable must be deleted. The existing system of bridges does not work, as it takes as input a MOI constraint (which, by design, the user-specified constraint have a low likelihood to be). Nothing is implemented regarding this. My idea is to have `_rewrite_expr` to return the new things that the function call created (variables, constraints), so that JuMP can get rid of all of it when deleting the constraint.