@dourouc05 2020-05-20
@blegat @odow @mlubin Is this approach viable? 

(Thinking about it, I suppose `_rewrite_expr` should lose its starting underscore, as it would be a part of the public interface.)