@dourouc05 2020-06-03
I just added a few tests for binary operators.