@dourouc05 2020-07-18
@odow @blegat Does this approach make sense? In other words, might the ideas behind this PR get merged? What do you think of the concept of one JuMP-level constraint mapped to several MOI constraints (say, `CompoundConstraint`)?