@dourouc05 2020-07-19
I did a little benchmark on this (https://github.com/dourouc05/JuMP.jl/blob/dourouc05-rewrite-complex/benchmark/parse_constraints.jl). Indeed, without writing complex things, this code may make parsing the constraints twice as slow. 

Before: 

```
julia> include("C:\\Users\\Thibaut\\.julia\\dev\\JuMP\\benchmark\\parse_constraints.jl")
(1/2) benchmarking "model0"...
done (took 5.4042206 seconds)
(2/2) benchmarking "model1"...
done (took 5.480212001 seconds)
2-element BenchmarkTools.BenchmarkGroup:
  tags: []
  "model0" => Trial(28.958 ms)
  "model1" => Trial(27.174 ms)
```

With this PR:

```
julia> include("C:\\Users\\Thibaut\\.julia\\dev\\JuMP\\benchmark\\parse_constraints.jl")
(1/2) benchmarking "model0"...
done (took 5.499017699 seconds)
(2/2) benchmarking "model1"...
done (took 5.5020427 seconds)
2-element BenchmarkTools.BenchmarkGroup:
  tags: []
  "model0" => Trial(29.167 ms)
  "model1" => Trial(55.927 ms)
```

While you don't look opposed to the idea, I guess the implementation is not performant enough. 

I'm not sure I get the details of the code you point at. You are generating "fast" code to parse the head of an expression, right? Plus using `Object` so that Julia doesn't generate too much code for different types?