@blegat 2020-07-19
> What do you think of the concept of one JuMP-level constraint mapped to several MOI constraints (say, `CompoundConstraint`)?

You could generate a single constraint that is bridged into several constraints. I'd prefer not having constraint returning several MOI constraints