@dourouc05 2020-07-19
> You could generate a single constraint that is bridged into several constraints. I'd prefer not having constraint returning several MOI constraints

Which "single" constraint? Does MOI need a way to represent constraints like `count(x .== 0) <= count(y .<= z) + minimum(abs.(x))`? That would require adding `count` as a new type of function, along with `minimum` and anything that might be used in an expression, but also a way to combine all these (at least in an affine way, but probably also nonlinear). It would make MOI's CP extension much much more complex. The solver interface would have more information about the constraints and could potentially use it (some CP solvers have a way to represent `x <= count(sth)` directly, but not all of them… and none can really gulp down more complex things), but implementing such an interface would be much more cumbersome (all solver interfaces would need a way to parse these expressions). 

That's why I think it would make sense, at the modelling layer, to decompose complex constraints into bite-sized MOI constraints, at the lowest level of abstraction that makes sense based on the solver APIs (that would mean adding things like `x <= count(sth)` in the MOI extensions, but nothing more complicated).