@odow 2020-07-20
> could be a solution to implement this pattern matching, with high performance, but no extension possibilities (which is the point, here).

Yes. This is the point I settled on. Create a giant `if` statement with the common ones first. Make it easy to extend if we add stuff to JuMP, but don't make it externally extensible.