@dourouc05 2020-07-20
> Yes. This is the point I settled on. Create a giant `if` statement with the common ones first. Make it easy to extend if we add stuff to JuMP, but don't make it externally extensible.

That would defeat the whole point of this, which is to allow extensions of JuMP syntax. Or maybe there is a way to keep a fast path for the common case, and keep the flexibility for extensions (with a performance penalty)?