@blegat 2020-07-20
It seems your use case is similar to disciplined convex reformulation. In JuMP, you need to write your convex model as constraints that are either `VectorOfVariables`-in-`Cone` or `VectorAffineFunction`-in-`Cone`.
However, if a nonlinear expression is disciplined convex (which is stronger than convex), there is an automatic way to rewrite it as constraints of the form just mentioned.
Convex.jl does this reformulation for instance.
The plan to bring this kind of reformulation in JuMP is to wait for https://github.com/jump-dev/MathOptInterface.jl/issues/846 and then to add bridges for nonlinear expression into convex constraints.
This is a bit different from other bridges as you cannot determine just from the type of the constraints whether you will be able to transform it into convex constraints and which constraints you are going to produce so we might need to adapt `LazyBridgeOptimizer` or apply this bridge with a `SingleBridgeOptimizer`.
You could have the same approach and bridge nonlinear expression into constraint programming sets in bridges instead of doing it during the parsing of the macros.