@dourouc05 2020-07-20
> It seems your use case is similar to disciplined convex reformulation. 

Indeed, it's the same issue, as DCP may decompose a single expression in many constraints. For now, I can live without the possibility to remove CP constraints :)! 

(By the way, I didn't see any documents giving ideas on how to implement bridges on top of these nonlinear expressions, that's why I didn't think this issue was relevant.)