@dourouc05 2020-07-20
@odow I've been benchmarking this code a tad more, and it looks like my previous figures where wrong (I was using Revise to test the latest version of the code). When restarting Julia between runs (i.e. with and without this PR), I cannot find a real performance difference. 

Before: 

```
julia> include("C:\\Users\\Thibaut\\.julia\\dev\\JuMP\\benchmark\\parse_constraints.jl")
(1/2) benchmarking "model0"...
done (took 5.4042206 seconds)
(2/2) benchmarking "model1"...
done (took 5.480212001 seconds)
2-element BenchmarkTools.BenchmarkGroup:
  tags: []
  "model0" => Trial(29.564 ms)
  "model1" => Trial(53.224 ms)
```

After: 

```
julia> include("C:\\Users\\Thibaut\\.julia\\dev\\JuMP\\benchmark\\parse_constraints.jl")
(1/2) benchmarking "model0"...
done (took 5.4042206 seconds)
(2/2) benchmarking "model1"...
done (took 5.480212001 seconds)
2-element BenchmarkTools.BenchmarkGroup:
  tags: []
  "model0" => Trial(28.770 ms)
  "model1" => Trial(53.410 ms)
```

I've run both quite a few times, and the figures never wander too far from these (give or take one or two milliseconds, for both). I've also tested to replace all these functions by a single one and a large `if`, but it gave no real difference.