@blegat 2020-07-22
I think the difference will be mainly on the compilation since it will need to compile many more methods due to `Val`