@dourouc05 2020-07-23
By the way, is performance really important? Most use cases I can think of should rather be implemented in nonlinear bridges rather than this way (everything, except overloading operators like `&&` between binary variables: maybe that should go into MutableArithmetics?). As I understand the new nonlinear implementation is still far away, it might be good to include this in the meantime.