@dourouc05 2020-08-20
@blegat @odow @mlubin As this code is intended to provide extension points for only a limited amount of time (until MOI fully supports nonlinearity), maybe it's in a good-enough shape to be merged? (At least as a 'private' API, i.e. not exported and potentially undocumented either.)