@blegat 2020-08-20
I'm a bit hesitant because it significantly increase compile time